+++ packages/core/theme/src/components/select.ts
@@ -22,7 +22,7 @@ const select = tv({
     ],
     mainWrapper: "w-full flex flex-col",
     trigger:
-      "relative px-3 gap-3 w-full inline-flex flex-row items-center shadow-xs outline-hidden tap-highlight-transparent",
+      "relative px-3 gap-3 w-full inline-flex flex-row items-center shadow-xs outline-solid outline-transparent tap-highlight-transparent",
     innerWrapper:
       "inline-flex h-fit w-[calc(100%_-theme(spacing.6))] min-h-4 items-center gap-1.5 box-border",
     selectorIcon: "absolute end-3 w-4 h-4",
@@ -31,9 +31,29 @@ const select = tv({
     listboxWrapper: "scroll-py-6 w-full",
     listbox: "",
     popoverContent: "w-full p-1 overflow-hidden",
+    clearButton: [
+      "w-4",
+      "h-4",
+      "z-10",
+      "mb-4",
+      "relative",
+      "start-auto",
+      "appearance-none",
+      "outline-none",
+      "select-none",
+      "opacity-70",
+      "hover:!opacity-100",
+      "cursor-pointer",
+      "active:!opacity-70",
+      "rounded-full",
+      // focus ring
+      ...dataFocusVisibleClasses,
+    ],
     helperWrapper: "p-1 flex relative flex-col gap-1.5 group-data-[has-helper=true]:flex",
     description: "text-tiny text-foreground-400",
     errorMessage: "text-tiny text-danger",
+    endWrapper: "flex end-18",
+    endContent: "mb-4",
   },
   variants: {
     variant: {
@@ -43,6 +63,7 @@ const select = tv({
           "data-[hover=true]:bg-default-200",
           "group-data-[focus=true]:bg-default-200",
         ],
+        clearButton: "mb-4",
       },
       faded: {
         trigger: [
@@ -52,6 +73,7 @@ const select = tv({
           "data-[hover=true]:border-default-400 data-[focus=true]:border-default-400 data-[open=true]:border-default-400",
         ],
         value: "group-data-[has-value=true]:text-default-foreground",
+        clearButton: "mb-4",
       },
       bordered: {
         trigger: [
@@ -62,6 +84,7 @@ const select = tv({
           "data-[focus=true]:border-default-foreground",
         ],
         value: "group-data-[has-value=true]:text-default-foreground",
+        clearButton: "mb-4",
       },
       underlined: {
         trigger: [
@@ -88,6 +111,7 @@ const select = tv({
           "data-[focus=true]:after:w-full",
         ],
         value: "group-data-[has-value=true]:text-default-foreground",
+        clearButton: "mb-4 me-2",
       },
     },
     color: {
@@ -113,14 +137,17 @@ const select = tv({
         label: "text-tiny",
         trigger: "h-8 min-h-8 px-2 rounded-small",
         value: "text-small",
+        clearButton: "text-medium",
       },
       md: {
         trigger: "h-10 min-h-10 rounded-medium",
         value: "text-small",
+        clearButton: "text-large",
       },
       lg: {
         trigger: "h-12 min-h-12 rounded-large",
         value: "text-medium",
+        clearButton: "mb-5 text-large",
       },
     },
     radius: {
@@ -143,10 +170,12 @@ const select = tv({
     labelPlacement: {
       outside: {
         base: "flex flex-col",
+        clearButton: "mb-0",
       },
       "outside-left": {
         base: "flex-row items-center flex-nowrap data-[has-helper=true]:items-start",
         label: "relative pe-2 text-foreground",
+        clearButton: "mb-0",
       },
       inside: {
         label: "text-tiny cursor-pointer",
@@ -159,6 +188,12 @@ const select = tv({
       },
       false: {
         base: "min-w-40",
+      },
+    },
+    isClearable: {
+      true: {
+        clearButton: "peer-data-[filled=true]:opacity-70 peer-data-[filled=true]:block",
+        endContent: "ms-3",
       },
     },
     isDisabled: {
@@ -208,6 +243,7 @@ const select = tv({
           "motion-reduce:transition-none",
         ],
         selectorIcon: "transition-transform duration-150 ease motion-reduce:transition-none",
+        clearButton: ["transition-opacity", "motion-reduce:transition-none"],
       },
     },
     disableSelectorIconRotation: {
@@ -727,7 +763,7 @@ const select = tv({
         base: "data-[has-label=true]:mt-[calc(var(--heroui-font-size-small)_+_12px)]",
       },
     },
-    // outside-left & size & hasHelper
+    // outside-left & size
     {
       labelPlacement: "outside-left",
       size: "sm",
@@ -762,6 +798,46 @@ const select = tv({
       labelPlacement: ["inside", "outside"],
       class: {
         label: ["pe-2", "max-w-full", "text-ellipsis", "overflow-hidden"],
+      },
+    },
+    // isClearable & labelPlacement
+    {
+      labelPlacement: ["outside", "outside-left"],
+      isClearable: true,
+      class: {
+        endContent: ["mt-4"],
+        clearButton: ["group-data-[has-end-content=true]:mt-4"],
+      },
+    },
+    {
+      isClearable: false,
+      labelPlacement: ["outside", "outside-left"],
+      class: {
+        endContent: ["mt-4"],
+      },
+    },
+    // isClearable + variant
+    {
+      isClearable: true,
+      variant: ["underlined"],
+      class: {
+        clearButton: ["relative group-data-[has-end-content=true]:left-2"],
+        endContent: ["me-2"],
+      },
+    },
+    {
+      isClearable: false,
+      variant: ["underlined"],
+      class: {
+        endContent: ["me-2"],
+      },
+    },
+    // isClearable + size
+    {
+      isClearable: true,
+      size: "sm",
+      class: {
+        endContent: "ms-2",
       },
     },
   ],