+++ packages/components/kbd/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @heroui/kbd
 
+## 2.2.18-beta.3
+
+### Patch Changes
+
+- [`1bca3f9`](https://github.com/heroui-inc/heroui/commit/1bca3f994655081f04714843047185aacdd481c0) Thanks [@wingkwong](https://github.com/wingkwong)! - sync 2.7.11 release
+
+- Updated dependencies [[`1bca3f9`](https://github.com/heroui-inc/heroui/commit/1bca3f994655081f04714843047185aacdd481c0)]:
+  - @heroui/shared-utils@2.1.10-beta.5
+  - @heroui/react-utils@2.1.12-beta.3
+  - @heroui/system-rsc@2.3.16-beta.3
+
 ## 2.2.18-beta.2
 
 ### Patch Changes